rename events.py to ics.py
